Add repository migration message to the README
Gonna push this to a special `repo-migration` branch for GitHub only,
then set that as the default branch, I think? I would prefer for
clarity to not have two diverging concepts of the `main` branch, but
that's not a big deal in practice if that turns out to be necessary!